Extract TableLayoutControlCollection from TableLayoutPanel.cs
