+++ src/System.Windows.Forms/src/System/Windows/Forms/TableLayoutPanel.cs
@@ -566,30 +566,4 @@ namespace System.Windows.Forms
 
         #endregion
     }
-
-    /// <summary>
-    ///  Represents a collection of controls on the TableLayoutPanel.
-    /// </summary>
-    [ListBindable(false)]
-    [DesignerSerializer("System.Windows.Forms.Design.TableLayoutControlCollectionCodeDomSerializer, " + AssemblyRef.SystemDesign, "System.ComponentModel.Design.Serialization.CodeDomSerializer, " + AssemblyRef.SystemDesign)]
-    public class TableLayoutControlCollection : Control.ControlCollection
-    {
-        public TableLayoutControlCollection(TableLayoutPanel container) : base(container)
-        {
-            Container = container ?? throw new ArgumentNullException(nameof(container));
-        }
-
-        //the container of this TableLayoutControlCollection
-        public TableLayoutPanel Container { get; }
-
-        /// <summary>
-        ///  Add control to cell (x, y) on the table. The control becomes absolutely positioned if neither x nor y is equal to -1
-        /// </summary>
-        public virtual void Add(Control control, int column, int row)
-        {
-            base.Add(control);
-            Container.SetColumn(control, column);
-            Container.SetRow(control, row);
-        }
-    }
 }
